+++ callbacks/settings-changelog.ts
@@ -1,7 +1,7 @@
 import { Context } from "grammy";
-import { trackEvent } from "../../helpers/analytics";
-import { updateSettings } from "../../helpers/api";
-import { changelogMessageTemplate } from "../../helpers/templates";
+import { trackEvent } from "../helpers/analytics";
+import { updateSettings } from "../helpers/api";
+import { changelogMessageTemplate } from "../helpers/templates";
 import { deleteMessage } from "../actions/delete-message";
 
 const FIELD_NAME = "changelog";
