+++ commands/permissions.ts
@@ -1,10 +1,11 @@
-import { bot } from "../..";
+import { bot } from "..";
 import { Context } from "grammy";
-import { trackEvent } from "../../helpers/analytics";
+import { trackEvent } from "../helpers/analytics";
 import { deleteMessage } from "../actions/delete-message";
 import { showBotActivity } from "../actions/show-bot-activity";
+import { handleMissingPermissions } from "../actions/missing-permissions";
 
-bot.command("source", async (ctx: Context) => {
+bot.command("permissions", async (ctx: Context) => {
   if (!ctx.msg) return;
 
   const chatId = ctx?.msg?.chat.id;
@@ -12,7 +13,7 @@ bot.command("source", async (ctx: Context) => {
 
   showBotActivity(chatId);
   deleteMessage(chatId, msgId);
-  ctx.reply(`This bot’s source code is available on GitHub: https://github.com/pugson/telegram-twitter-url-expand-bot`);
+  handleMissingPermissions(ctx, true);
 
-  trackEvent("command.sourceCode");
+  trackEvent("command.permissions");
 });
