--- handlers/callbacks/expand-choice.ts
@@ -1,2 +1,0 @@
-// Expand this X?
-// [ Yes ] [ No ]
